refactor: separa responsabilidade de login/autenticação para um modulo dedicado
