+++ api/app/modules/autenticacao/autenticacao_schema.py
@@ -1,0 +1,11 @@
+from pydantic import BaseModel, EmailStr
+
+
+class LoginSchema(BaseModel):
+    email: EmailStr
+    senha: str
+
+
+class TokenSchema(BaseModel):
+    access_token: str
+    token_type: str = "bearer"
