+++ api/app/modules/usuario/usuario_schema.py
@@ -92,16 +92,3 @@ class UsuarioResponse(UsuarioBase):
     id: int
     status: str
     perfis: list[str] = []
-
-
-# --- Autenticação ---
-
-
-class LoginSchema(BaseModel):
-    email: EmailStr
-    senha: str
-
-
-class TokenSchema(BaseModel):
-    access_token: str
-    token_type: str = "bearer"
